+++ src/main/java/racingcar/model/position/History.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.model.position;
 
 import java.util.ArrayList;
 import java.util.Collections;
