+++ src/main/java/racingcar/model/position/Position.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.model.position;
 
 import java.util.Objects;
