+++ src/main/java/racingcar/util/comparable/NumberComparable.java
@@ -1,4 +1,4 @@
-package racingcar.comparable;
+package racingcar.util.comparable;
 
 public interface NumberComparable<T extends Number> {
     int compare(T x, T y);
