+++ src/main/java/racingcar/util/random/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.random;
+package racingcar.util.random;
 
 public interface RandomNumberGenerator<T extends Number> {
     T pickNumber();
