+++ src/main/java/racingcar/util/splitter/Splitter.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.util.splitter;
 
 public class Splitter {
     public String[] splitFrom(final String inputNames, final String delimiter) {
