+++ src/main/java/racingcar/view/ConsoleInputHandler.java
@@ -1,4 +1,4 @@
-package racingcar.io;
+package racingcar.view;
 
 import camp.nextstep.edu.missionutils.Console;
