+++ src/main/java/racingcar/view/InputHandler.java
@@ -1,4 +1,4 @@
-package racingcar.io;
+package racingcar.view;
 
 public interface InputHandler {
     String read();
